requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.1
django-redis==5.4.0
idna==3.6
iniconfig==2.0.0
mysqlclient==2.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pluggy==1.4.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.2
tzdata==2023.4
urllib3==2.2.0
ximilar-client==1.30.5
